debug base class

+++ tmpsrc/SlaveConfig.cpp
@@ -6,7 +6,7 @@ int getSlaveIntConf(std::string key)
    return confValue; 
 }
 
-double getSlaveFloatConf(std::string key)
+float getSlaveFloatConf(std::string key)
 {
    float confValue = cf.Value(SlaveKey, key);
    return confValue; 
@@ -15,5 +15,5 @@ double getSlaveFloatConf(std::string key)
 std::string getSlaveStringConf(std::string key)
 {
     std::string confValue = cf.Value(SlaveKey, key);
-   return confValue;
+   return confValue; 
 }
